+++ app/Http/Controllers/SubCategoryController.php
@@ -6,6 +6,8 @@ use Illuminate\Http\Request;
 use App\Models\SubCategory;
 use App\Models\Category;
 use Yajra\Datatables\Datatables;
+use Intervention\Image\Facades\Image as Image;
+
 class SubCategoryController extends Controller
 {
     public function __construct()
@@ -30,9 +32,12 @@ class SubCategoryController extends Controller
             $name= time().'.'.$name_array[0];
             $file->move(public_path('images/assets'), $name);
             $input['image']='/images/assets/'.$name;
+            $image=Image::make(public_path($input['image']))->resize(540, 300);
+            $image->save(public_path('images/assets/compressed_'.$name));
+            $input['compressed_image']='/images/assets/'.$name;
         }
         SubCategory::create($input);
-        return redirect()->back();
+        return redirect()->back()->with('message', 'Record Added successfully!!!');
     }
     public function edit_sub_category(Request $request)
     {
@@ -45,10 +50,16 @@ class SubCategoryController extends Controller
        		if(file_exists(public_path($cat->image))) {
             unlink(public_path($cat->image));
         }
+        if(file_exists(public_path($cat->compressed_image))) {
+            unlink(public_path($cat->compressed_image));
+        }
             $name_array=array_map('strrev', explode('.', strrev($file->getClientOriginalName())));   
             $name= time().'.'.$name_array[0];
             $file->move(public_path('images/assets'), $name);
             $input['image']='/images/assets/'.$name;
+            $image=Image::make(public_path($input['image']))->resize(540, 300);
+            $image->save(public_path('images/assets/compressed_'.$name));
+            $input['compressed_image']='/images/assets/'.$name;
         }
         $input['name']=$request->name;
         $input['slug']=$request->slug;
@@ -59,11 +70,17 @@ class SubCategoryController extends Controller
     }
      public function delete_sub_category($id){
     		$cat=SubCategory::find($id);
+            if(count($cat->third_category) > 0 ||  count($cat->products) > 0){
+                return redirect()->back()->withErrors(['This category can\'t be deleted, because it contains Sub Category or products']);
+            }
     		if(file_exists(public_path($cat->image))) {
             unlink(public_path($cat->image));
         	}
+            if(file_exists(public_path($cat->compressed_image))) {
+                unlink(public_path($cat->compressed_image));
+            }
     		$cat->delete();
-    		return redirect()->back();
+    		return redirect()->back()->with('message', 'Record deleted successfully!!!');
     }	
     public function get_sub_category(Request $request){
         return SubCategory::where('category_id',$request->parent_id)->get();
